Move to_nova_server to a classmethod on NovaServer
