+++ otter/convergence/gathering.py
@@ -4,7 +4,6 @@ from urllib import urlencode
 from effect import parallel
 
 from toolz.curried import filter, groupby, keyfilter, map
-from toolz.dicttoolz import get_in
 from toolz.functoolz import compose, identity
 
 from otter.constants import ServiceType
@@ -171,39 +170,6 @@ def extract_CLB_drained_at(feed):
         raise ValueError('Unexpected summary: {}'.format(summary))
 
 
-def _private_ipv4_addresses(server):
-    """
-    Get all private IPv4 addresses from the addresses section of a server.
-
-    :param dict server: A server dict.
-    :return: List of IP addresses as strings.
-    """
-    private_addresses = get_in(["addresses", "private"], server, [])
-    return [addr['addr'] for addr in private_addresses if addr['version'] == 4]
-
-
-def _servicenet_address(server):
-    """
-    Find the ServiceNet address for the given server.
-    """
-    return next((ip for ip in _private_ipv4_addresses(server)
-                 if ip.startswith("10.")), "")
-
-
-def to_nova_server(server_json):
-    """
-    Convert from JSON format to :obj:`NovaServer` instance.
-    """
-    return NovaServer(id=server_json['id'],
-                      state=ServerState.lookupByName(server_json['status']),
-                      created=timestamp_to_epoch(server_json['created']),
-                      image_id=server_json.get('image', {}).get('id'),
-                      flavor_id=server_json['flavor']['id'],
-                      desired_lbs=NovaServer.lbs_from_metadata(
-                          server_json.get('metadata')),
-                      servicenet_address=_servicenet_address(server_json))
-
-
 def get_all_convergence_data(
         group_id,
         get_scaling_group_servers=get_scaling_group_servers,
@@ -217,7 +183,7 @@ def get_all_convergence_data(
     eff = parallel(
         [get_scaling_group_servers()
          .on(lambda servers: servers.get(group_id, []))
-         .on(map(to_nova_server)).on(list),
+         .on(map(NovaServer.from_server_details_json)).on(list),
          get_clb_contents()]
     ).on(tuple)
     return eff
